Make uint32_t and int32_t definitions align with stdint definitions.

+++ phase2_embedded_c/globalDefs.h
@@ -22,11 +22,11 @@
 
 typedef unsigned char  uint8_t;
 typedef unsigned short uint16_t;
-typedef unsigned int   uint32_t;
+typedef unsigned long  uint32_t;
 
 typedef signed char    int8_t;
 typedef signed short   int16_t;
-typedef signed int     int32_t;
+typedef signed long    int32_t;
 
 typedef unsigned char  bool8_t;
 typedef unsigned short bool16_t;
